requirements.txt
Django
djangorestframework
drf-yasg
djangorestframework-simplejwt
psycopg2
redis
